Code now using poetry and alembic.
Changed the code to be easily managed using poetry and alembic. Dependencies can now be easily managed, and alembic takes care of the initial creation of tables and future updates in/to the database. Dependencies and installation notes are more detailed. This makes it easier for someone to install the services in another system.